Removed base list item banner.
