+++ android/app/src/main/java/com/koleff/kare_android/ui/compose/banners/ExerciseBanner.kt
@@ -43,6 +43,7 @@ import androidx.compose.ui.unit.sp
 import androidx.navigation.NavHostController
 import androidx.navigation.compose.rememberNavController
 import com.koleff.kare_android.R
+import com.koleff.kare_android.data.MainScreen
 import com.koleff.kare_android.data.model.dto.ExerciseDto
 import com.koleff.kare_android.data.model.dto.MachineType
 import com.koleff.kare_android.data.model.dto.MuscleGroup
@@ -243,6 +244,15 @@ fun ExerciseBannerV2(
             }
         }
     }
+}
+
+fun openExerciseDetailsScreen(exercise: ExerciseDto, navController: NavHostController) {
+    navController.navigate(
+        MainScreen.ExerciseDetails.createRoute(
+            exerciseId = exercise.exerciseId,
+            muscleGroupId = exercise.muscleGroup.muscleGroupId
+        )
+    )
 }
 
 @Composable
